refactor: consolidate examples from 15 to 8 files
- Merge parallel_progress_demo, tree_progress_demo, simple_progress_demo,
  and system_call_demo into unified progress_demo.rb
- Remove redundant examples: nested_section_demo, data_pipeline_demo,
  large_tree_demo (overlapping with existing demos)
- Add plain progress mode documentation to README and docs/guide.md
- Add Lifecycle Management section (clean/run_and_clean) to README
- Update examples/README.md with new structure
